requirements.txt
fastapi
python-dotenv 

uvicorn

sqlalchemy
psycopg2-binary